Add Alloc Speed Test
The test is inconclusive. I am still making the PR in order to discuss
possible solutions.

The allocation times fluctuate from 1.90735e-06 to ~2.5 sec.
I am not sure if this is because of the OS memory load by the time of
the execution of if this is related to the small memory allocation of
the test.